new assessment status middleware.
conditionally display notification on each page.
