+++ server/views/macros/incompleteBCST2.njk
@@ -1,0 +1,17 @@
+{% macro incompleteBCST2(pathwayName, prisonerData) %}
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+      <h2 class="govuk-heading-l">{{ pathwayName }}</h2>
+      <div class="govuk-notification-banner" role="region" aria-labelledby="govuk-notification-banner-title" data-module="govuk-notification-banner">
+        <div class="govuk-notification-banner__header">
+          <h2 class="govuk-notification-banner__title" id="govuk-notification-banner-title">Important</h2>
+        </div>
+        <div class="govuk-notification-banner__content">
+          <p class="govuk-notification-banner__heading">{{ prisonerData.personalDetails.firstName | capitalize }} {{ prisonerData.personalDetails.lastName | capitalize }} does not have a completed BCST2 assessment.<br /><br />
+            <a class="govuk-notification-banner__link" href="/assessment-task-list/?prisonerNumber={{ prisonerData.personalDetails.prisonerNumber }}">Complete assessment now</a>
+          </p>
+        </div>
+      </div>
+    </div>
+  </div>
+{% endmacro %}
