+++ server/views/pages/accommodation.njk
@@ -2,102 +2,105 @@
 {% set pathway = "Accommodation" %}
 {% set pageTitle = applicationName + " - " + pathway %}
 {% set pathwayStatus = prisonerData.pathways | filterByPathway( pathway | getEnumByName ) %}
+{% set resettlementAssessmentEnabled = features.RESETTLEMENT_ASSESSMENT | getFeatureFlag %}
 
 {% set mainClasses = "app-container govuk-body" %}
 {% block content %}
-  <div class="govuk-grid-row govuk-!-padding-top-4 govuk-!-padding-bottom-4">
-    {{ titleAndIntro(pathway) }}
-    <div
-        class="govuk-grid-column-one-third">
-      <!-- Pathway status -->
-      {{ statusDetails(pathwayStatus, prisonerData.personalDetails.prisonerNumber, pathway) }}
+  {% if not resettlementAssessmentEnabled or (resettlementAssessmentEnabled and BCST2Completed) %}
+    <div class="govuk-grid-row govuk-!-padding-top-4 govuk-!-padding-bottom-4">
+      {{ titleAndIntro(pathway) }}
+      <div class="govuk-grid-column-one-third">
+        {{ statusDetails(pathwayStatus, prisonerData.personalDetails.prisonerNumber, pathway) }}
+      </div>
     </div>
-  </div>
-  <div class="govuk-grid-row govuk-!-padding-top-4">
-    <div class="govuk-grid-column-one-quarter sticky-anchor-links">
-      {% set subNavItems = [
-        {
-          name: "Duty to refer",
-          id: "duty-to-refer"
-        },
-        {
-          name: "Main address",
-          id: "main-address"
-        }
-      ] %}
-      {% set subNavItems = (subNavItems.concat(crsReferrals.results[0].referrals | createReferralsSubNav)) %}
-      {{ subNav(subNavItems) }}
-    </div>
-    <div class="govuk-grid-column-three-quarters">
-      <section id="duty-to-refer" class="app-summary-card govuk-!-margin-bottom-8">
-        <header class="app-summary-card__header">
-          <h3 class="app-summary-card__title">
-            Duty to refer
-            <span class="govuk-caption-m">Data sourced from nDelius</span>
-          </h3>
-          <span class="right"></span>
-        </header>
-        <div class="app-summary-card__body">
-          {% if accommodation.status %}
-            {{ govukTable({
-              firstCellIsHeader: true,
-              rows: [
-                [
-                  {
-                    text: "Referral date"
-                  },
-                  {
-                    text: accommodation.referralDate | formatDate('long')
-                  }
-                ],
-                [
-                  {
-                    text: "Status"
-                  },
-                  {
-                    text: accommodation.status
-                  }
+    <div class="govuk-grid-row govuk-!-padding-top-4">
+      <div class="govuk-grid-column-one-quarter sticky-anchor-links">
+        {% set subNavItems = [
+          {
+            name: "Duty to refer",
+            id: "duty-to-refer"
+          },
+          {
+            name: "Main address",
+            id: "main-address"
+          }
+        ] %}
+        {% set subNavItems = (subNavItems.concat(crsReferrals.results[0].referrals | createReferralsSubNav)) %}
+        {{ subNav(subNavItems) }}
+      </div>
+      <div class="govuk-grid-column-three-quarters">
+        <section id="duty-to-refer" class="app-summary-card govuk-!-margin-bottom-8">
+          <header class="app-summary-card__header">
+            <h3 class="app-summary-card__title">
+              Duty to refer
+              <span class="govuk-caption-m">Data sourced from nDelius</span>
+            </h3>
+            <span class="right"></span>
+          </header>
+          <div class="app-summary-card__body">
+            {% if accommodation.status %}
+              {{ govukTable({
+                firstCellIsHeader: true,
+                rows: [
+                  [
+                    {
+                      text: "Referral date"
+                    },
+                    {
+                      text: accommodation.referralDate | formatDate('long')
+                    }
+                  ],
+                  [
+                    {
+                      text: "Status"
+                    },
+                    {
+                      text: accommodation.status
+                    }
+                  ]
                 ]
-              ]
-            }) }}
-          {% elseif not accommodation.error %}
-            No duty to refer details found
-          {% else %}
-            {{ accommodation.error }}
-          {% endif %}
-        </div>
-      </section>
-      <section id="main-address" class="app-summary-card govuk-!-margin-bottom-8">
-        <header class="app-summary-card__header">
-          <h3 class="app-summary-card__title">
-            Main address
-            <span class="govuk-caption-m">Data sourced from nDelius</span>
-          </h3>
-          <span class="right"></span>
-        </header>
-        <div class="app-summary-card__body">
-          {% if accommodation.message %}
-            <p>{{ accommodation.message }}</p>
-          {% elseif not accommodation.error %}
-            {{ govukTable({
-              firstCellIsHeader: true,
-              rows: [
-                [
-                  {
-                    text: "Address"
-                  },
-                  {
-                    text: accommodation.mainAddress
-                  }
+              }) }}
+            {% elseif not accommodation.error %}
+              No duty to refer details found
+            {% else %}
+              {{ accommodation.error }}
+            {% endif %}
+          </div>
+        </section>
+        <section id="main-address" class="app-summary-card govuk-!-margin-bottom-8">
+          <header class="app-summary-card__header">
+            <h3 class="app-summary-card__title">
+              Main address
+              <span class="govuk-caption-m">Data sourced from nDelius</span>
+            </h3>
+            <span class="right"></span>
+          </header>
+          <div class="app-summary-card__body">
+            {% if accommodation.message %}
+              <p>{{ accommodation.message }}</p>
+            {% elseif not accommodation.error %}
+              {{ govukTable({
+                firstCellIsHeader: true,
+                rows: [
+                  [
+                    {
+                      text: "Address"
+                    },
+                    {
+                      text: accommodation.mainAddress
+                    }
+                  ]
                 ]
-              ]
-            }) }}
-          {% else %}
-            {{ accommodation.error }}
-          {% endif %}
-        </div>
-      </section>
-      {% include "../partials/crsReferrals.njk" %}
+              }) }}
+            {% else %}
+              {{ accommodation.error }}
+            {% endif %}
+          </div>
+        </section>
+        {% include "../partials/crsReferrals.njk" %}
+      </div>
     </div>
-  </div>
+  {% else %}
+    {{ incompleteBCST2(pathway, prisonerData) }}
+  {% endif %}
 {% endblock %}
