+++ server/views/pages/children-families-communities.njk
@@ -3,23 +3,25 @@
 {% set pageTitle = applicationName + " - " + pathway %}
 {% set mainClasses = "app-container govuk-body" %}
 {% set pathwayStatus = prisonerData.pathways | filterByPathway( pathway | getEnumByName ) %}
+{% set resettlementAssessmentEnabled = features.RESETTLEMENT_ASSESSMENT | getFeatureFlag %}
+
 {% block content %}
+  {% if not resettlementAssessmentEnabled or (resettlementAssessmentEnabled and BCST2Completed) %}
     <div class="govuk-grid-row govuk-!-padding-top-4 govuk-!-padding-bottom-4">
-        {{ titleAndIntro(pathway) }}
-        <div
-                class="govuk-grid-column-one-third">
-            <!-- Pathway status -->
-            {{ statusDetails(pathwayStatus, prisonerData.personalDetails.prisonerNumber, pathway) }}
-        </div>
+      {{ titleAndIntro(pathway) }}
+      <div class="govuk-grid-column-one-third">
+        {{ statusDetails(pathwayStatus, prisonerData.personalDetails.prisonerNumber, pathway) }}
+      </div>
     </div>
-    <div
-            class="govuk-grid-row govuk-!-padding-top-4">
-        <!-- Main content area -->
-        <div class="govuk-grid-column-one-quarter sticky-anchor-links">
-            {{ subNav(crsReferrals.results[0].referrals | createReferralsSubNav) }}
-        </div>
-        <div class="govuk-grid-column-three-quarters">
-            {% include "../partials/crsReferrals.njk" %}
-        </div>
+    <div class="govuk-grid-row govuk-!-padding-top-4">
+      <div class="govuk-grid-column-one-quarter sticky-anchor-links">
+        {{ subNav(crsReferrals.results[0].referrals | createReferralsSubNav) }}
+      </div>
+      <div class="govuk-grid-column-three-quarters">
+        {% include "../partials/crsReferrals.njk" %}
+      </div>
     </div>
+  {% else %}
+    {{ incompleteBCST2(pathway, prisonerData) }}
+  {% endif %}
 {% endblock %}
