+++ server/views/pages/drugs-alcohol.njk
@@ -3,24 +3,27 @@
 {% set pageTitle = applicationName + " - " + pathway %}
 {% set mainClasses = "app-container govuk-body" %}
 {% set pathwayStatus = prisonerData.pathways | filterByPathway("DRUGS_AND_ALCOHOL") %}
+{% set resettlementAssessmentEnabled = features.RESETTLEMENT_ASSESSMENT | getFeatureFlag %}
+
 {% block content %}
-  <div class="govuk-grid-row govuk-!-padding-top-4 govuk-!-padding-bottom-4">
-    {{ titleAndIntro(pathway) }}
-    <!-- Pathway status -->
-    <div
-        class="govuk-grid-column-one-third">
-      <!-- Pathway status -->
-      {{ statusDetails(pathwayStatus, prisonerData.personalDetails.prisonerNumber, pathway) }}
+  {% if not resettlementAssessmentEnabled or (resettlementAssessmentEnabled and BCST2Completed) %}
+    <div class="govuk-grid-row govuk-!-padding-top-4 govuk-!-padding-bottom-4">
+      {{ titleAndIntro(pathway) }}
+      <div class="govuk-grid-column-one-third">
+        {{ statusDetails(pathwayStatus, prisonerData.personalDetails.prisonerNumber, pathway) }}
+      </div>
     </div>
-  </div>
-  <div
-      class="govuk-grid-row govuk-!-padding-top-4">
-    <!-- Main content area -->
-    <div class="govuk-grid-column-one-quarter sticky-anchor-links">
-      {{ subNav(crsReferrals.results[0].referrals | createReferralsSubNav) }}
+    <div class="govuk-grid-row govuk-!-padding-top-4">
+      <div class="govuk-grid-column-one-quarter sticky-anchor-links">
+        {{ subNav(crsReferrals.results[0].referrals | createReferralsSubNav) }}
+      </div>
+      <div class="govuk-grid-column-three-quarters">
+        {% if not resettlementAssessmentEnabled or (resettlementAssessmentEnabled and BCST2Completed) %}
+          {% include "../partials/crsReferrals.njk" %}
+        {% endif %}
+      </div>
     </div>
-    <div class="govuk-grid-column-three-quarters">
-      {% include "../partials/crsReferrals.njk" %}
-    </div>
-  </div>
+  {% else %}
+    {{ incompleteBCST2(pathway, prisonerData) }}
+  {% endif %}
 {% endblock %}
