+++ server/views/partials/layout.njk
@@ -4,6 +4,7 @@
 {% from "govuk/components/phase-banner/macro.njk" import govukPhaseBanner %}
 {% from "govuk/components/table/macro.njk" import govukTable %}
 {% from "../macros/title-and-intro.njk" import titleAndIntro %}
+{% from "../macros/incompleteBCST2.njk" import incompleteBCST2 %}
 {% from "../macros/status-details.njk" import statusDetails %}
 {% from "../macros/pathway-subnav.njk" import subNav %}
 {% from "govuk/components/phase-banner/macro.njk" import govukPhaseBanner %}
